+++ src/Microsoft.DotNet.Darc/src/Darc/Properties/launchSettings.json
@@ -2,8 +2,7 @@
   "profiles": {
     "Microsoft.DotNet.Darc": {
       "commandName": "Project",
-      "commandLineArgs": "verify",
-      "workingDirectory": "D:\\enlistments\\corefx"
+      "commandLineArgs": "authenticate"
     }
   }
 }
